+++ keyring/backends/SecretService.py
@@ -52,7 +52,7 @@ class Keyring(KeyringBackend):
             "service": service,
             "username": username
             }
-        label = "%s @ %s" % (username, service)
+        label = "Password for '%s' on '%s'" % (username, service)
         collection.create_item(label, attributes, password, replace=True)
 
     def delete_password(self, service, username):